ajuste de directorios, ideas de funcionamiento
